throw error if both claims_namespace_path and claims_namespace are set
